+++ src/lsp/main.ts
@@ -238,8 +238,9 @@ async function activate(context: vscode.ExtensionContext): Promise<void> {
     const extensionPath = context.extensionPath;
     const currentVersion = readVersionFile(extensionPath);
     let clojureLspPath = getClojureLspPath(extensionPath, util.isWindows);
-    if (currentVersion !== config.CLOJURE_LSP_VERSION) {
-        const downloadPromise = downloadClojureLsp(context.extensionPath, config.CLOJURE_LSP_VERSION);
+    const configuredVersion: string = config.getConfig().clojureLspVersion;
+    if (currentVersion !== configuredVersion) {
+        const downloadPromise = downloadClojureLsp(context.extensionPath, configuredVersion);
         vscode.window.setStatusBarMessage('$(sync~spin) Downloading clojure-lsp', downloadPromise);
         clojureLspPath = await downloadPromise;
     }